# src/KrisLange.UrlShortener/Models/NewUrlSpec.cs
/*
 * Link Shortening Service
 *
 * This is an API for a Link Shortening Service
 *
 * OpenAPI spec version: 1.0.0
 * Contact: kris@krislange.com
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */

using System;
using System.Linq;
using System.IO;
using System.Text;
using System.Collections;
using System.Collections.Generic;
using System.Collections.ObjectModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;

namespace KrisLange.UrlShortener.Models
{ 
    /// <summary>
    /// 
    /// </summary>
    [DataContract]
    public partial class NewUrlSpec : IEquatable<NewUrlSpec>
    { 
        /// <summary>
        /// Gets or Sets LongUrl
        /// </summary>
        [Required]
        
        [DataMember(Name="longUrl")]
        public string LongUrl { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class NewUrlSpec {\n");
            sb.Append("  LongUrl: ").Append(LongUrl).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public string ToJson()
        {
            return JsonConvert.SerializeObject(this, Formatting.Indented);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="obj">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object obj)
        {
            if (ReferenceEquals(null, obj)) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((NewUrlSpec)obj);
        }

        /// <summary>
        /// Returns true if NewUrlSpec instances are equal
        /// </summary>
        /// <param name="other">Instance of NewUrlSpec to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(NewUrlSpec other)
        {
            if (ReferenceEquals(null, other)) return false;
            if (ReferenceEquals(this, other)) return true;

            return 
                (
                    LongUrl == other.LongUrl ||
                    LongUrl != null &&
                    LongUrl.Equals(other.LongUrl)
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                var hashCode = 41;
                // Suitable nullity checks etc, of course :)
                    if (LongUrl != null)
                    hashCode = hashCode * 59 + LongUrl.GetHashCode();
                return hashCode;
            }
        }

        #region Operators
        #pragma warning disable 1591

        public static bool operator ==(NewUrlSpec left, NewUrlSpec right)
        {
            return Equals(left, right);
        }

        public static bool operator !=(NewUrlSpec left, NewUrlSpec right)
        {
            return !Equals(left, right);
        }

        #pragma warning restore 1591
        #endregion Operators
    }
}